[UPDATE] Refactor follow solid commitment
