+++ src/components/ExamDetailSegment/index.js
